合并DragonBonesResource和DragonBonesFactory，将DragonBonesFactory直接作为资源
